cleaned Level class

+++ MallowsVsDango/Classes/Towers/Cutter.cpp
@@ -90,7 +90,7 @@ void Cutter::attack(){
 	for (auto& cTarget : otherTargets){
 		if (cTarget != nullptr){
 			Slash* slash = nullptr;
-			if (level >= (int)getConfig()["cost"].size() - 1 || state == LIMIT_BURSTING) {
+			if (level >= getConfig()["cost"].size() - 1 || state == LIMIT_BURSTING) {
 				slash = DeepSlash::create(cTarget, settings->getDamage(level), getConfig()["deep_wound_duration"].asDouble(),
 					getConfig()["deep_wound_percent"].asDouble());
 			}
